requirements.txt
Flask
cryptography
mnemonic
Werkzeug
Flask-Babel
pyopenssl
flask-socketio
request
pycryptodome
sqlalchemy
python-telegram-bot
asyncio
Flask[async]


